Break out parseLocales
